dockeranalyser-ui/requirements.txt
cached-property==1.4.2
certifi==2018.4.16
chardet==3.0.4
distlib==0.2.7
Django==2.0.6
django-cors-headers==2.2.0
docker==3.3.0
docker-compose==1.21.2
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
idna==2.6
jsonschema==2.6.0
pytz==2018.4
PyYAML==3.12
requests==2.18.4
six==1.11.0
texttable==0.9.1
urllib3==1.22
websocket-client==0.48.0
